|Amey|Mod|Notification save and get endpoint|

+++ src/main/java/com/lti/mod/services/notificationservice/repository/NotificationRepository.java
@@ -4,8 +4,8 @@ import com.lti.mod.services.notificationservice.model.NotificationDO;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface NotificationRepository extends CrudRepository<NotificationDO,Long> {
+public interface NotificationRepository extends CrudRepository<NotificationDO,Integer> {
 
     @Query("SELECT u FROM NotificationDO u WHERE u.userId = ?1")
-    public NotificationDO getNotificationByUserId(Long userId);
+    public NotificationDO getNotificationByUserId(Integer userId);
 }
